shared/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
motor
fastapi
slowapi
